move the Scheme trait and add doc comments with examples
