+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "poprf"
-version = "0.1.8"
+version = "0.1.9"
 edition = "2021"
 
 description = "A threshold-computable partially-oblivous psuedo-random function"
